+++ core/runtime/chat_security.js
@@ -157,3 +157,6 @@ function createChatSecurity({ bot, pgPool }) {
 module.exports = { createChatSecurity, formatChatLine };
 
 
+
+
+